Katip.Core

Description

This module is not meant to be imported directly and may contain internal mechanisms that will change without notice.

Documentation

readMay :: Read a => String -> Maybe a

newtype Namespace

Represents a heirarchy of namespaces going from general to specific. For instance: ["processname", "subsystem"]. Note that single-segment namespaces can be created using IsString/OverloadedStrings, so "foo" will result in Namespace ["foo"].

Constructors

Namespace

Fields

unNamespace :: [Text]

intercalateNs :: Namespace -> [Text]

Ready namespace for emission with dots to join the segments.

newtype Environment

Application environment, like prod, devel, testing.

Constructors

Environment

Fields

getEnvironment :: Text

data Severity

Constructors

DebugS
Debug messages

InfoS
Information

NoticeS
Normal runtime Conditions

WarningS
General Warnings

ErrorS
General Errors

CriticalS
Severe situations

AlertS
Take immediate action

EmergencyS
System is unusable

data Verbosity

Verbosity controls the amount of information (columns) a Scribe emits during logging.

The convention is: - V0 implies no additional payload information is included in message. - V3 implies the maximum amount of payload information. - Anything in between is left to the discretion of the developer.

Constructors

V0

V1

V2

V3

renderSeverity :: Severity -> Text

textToSeverity :: Text -> Maybe Severity

newtype LogStr

Log message with Builder underneath; use <> to concat in O(1).

Constructors

LogStr

Fields

unLogStr :: Builder

logStr :: StringConv a Text => a -> LogStr

Pack any string-like thing into a LogStr. This will automatically work on String, ByteString, Text and any of the lazy variants.

ls :: StringConv a Text => a -> LogStr

Shorthand for logStr

showLS :: Show a => a -> LogStr

Convert any showable type into a LogStr.

newtype ThreadIdText

Constructors

ThreadIdText

Fields

getThreadIdText :: Text

mkThreadIdText :: ThreadId -> ThreadIdText

data Item a

This has everything each log message will contain.

Constructors

Item

Fields

_itemApp :: Namespace

_itemEnv :: Environment

_itemSeverity :: Severity

_itemThread :: ThreadIdText

_itemHost :: HostName

_itemProcess :: ProcessID

_itemPayload :: a

_itemMessage :: LogStr

_itemTime :: UTCTime

_itemNamespace :: Namespace

_itemLoc :: Maybe Loc

itemTime :: forall a. Lens' (Item a) UTCTime

itemThread :: forall a. Lens' (Item a) ThreadIdText

itemSeverity :: forall a. Lens' (Item a) Severity

itemProcess :: forall a. Lens' (Item a) ProcessID

itemPayload :: forall a a. Lens (Item a) (Item a) a a

itemNamespace :: forall a. Lens' (Item a) Namespace

itemMessage :: forall a. Lens' (Item a) LogStr

itemLoc :: forall a. Lens' (Item a) (Maybe Loc)

itemHost :: forall a. Lens' (Item a) HostName

itemEnv :: forall a. Lens' (Item a) Environment

itemApp :: forall a. Lens' (Item a) Namespace

newtype LocShow

Constructors

LocShow Loc

newtype LocJs

Constructors

LocJs

Fields

getLocJs :: Loc

processIDToText :: ProcessID -> Text

textToProcessID :: Text -> Maybe ProcessID

newtype ProcessIDJs

Constructors

ProcessIDJs

Fields

getProcessIDJs :: ProcessID

data PayloadSelection

Field selector by verbosity within JSON payload.

Constructors

AllKeys

SomeKeys [Text]

equivalentPayloadSelection :: PayloadSelection -> PayloadSelection -> Bool

Compares two payload selections for equivalence. With SomeKeys, ordering and duplicates are ignored.

class ToObject a where

Katip requires JSON objects to be logged as context. This typeclass provides a default instance which uses ToJSON and produces an empty object if toJSON results in any type other than object. If you have a type you want to log that produces an Array or Number for example, you'll want to write an explicit instance here. You can trivially add a ToObject instance for something with a ToJSON instance like:

instance ToObject Foo

Minimal complete definition

Nothing

Methods

toObject :: a -> Object

default toObject :: ToJSON a => a -> Object

class ToObject a => LogItem a where

Payload objects need instances of this class. LogItem makes it so that you can have very verbose items getting logged with lots of extra fields but under normal circumstances, if your scribe is configured for a lower verbosity level, it will only log a selection of those keys. Furthermore, each Scribe can be configured with a different Verbosity level. You could even use registerScribe, unregisterScribe, and clearScribes to at runtime swap out your existing scribes for more verbose debugging scribes if you wanted to.

When defining payloadKeys, don't redundantly declare the same keys for higher levels of verbosity. Each level of verbosity automatically and recursively contains all keys from the level before it.

Methods

payloadKeys :: Verbosity -> a -> PayloadSelection

List of keys in the JSON object that should be included in message.

data AnyLogPayload

Constructors

forall a.ToJSON a => AnyLogPayload a

newtype SimpleLogPayload

Constructors

SimpleLogPayload

Fields

unSimpleLogPayload :: [(Text, AnyLogPayload)]

toKey :: (Text, c) -> (Key, c)

sl :: ToJSON a => Text -> a -> SimpleLogPayload

Construct a simple log from any JSON item.

payloadObject :: LogItem a => Verbosity -> a -> Object

Constrain payload based on verbosity. Backends should use this to automatically bubble higher verbosity levels to lower ones.

filterElems :: [Text] -> KeyMap v -> KeyMap v

itemJson :: LogItem a => Verbosity -> Item a -> Value

Convert log item to its JSON representation while trimming its payload based on the desired verbosity. Backends that push JSON messages should use this to obtain their payload.

type PermitFunc = forall a. Item a -> IO Bool

Scribes are handlers of incoming items. Each registered scribe knows how to push a log item somewhere.

Guidelines for writing your own Scribe

Scribes should always take a Severity and Verbosity.

Severity is used to exclude log messages that are lower than the provided Severity. For instance, if the user passes InfoS, DebugS items should be ignored. Katip provides the permitItem utility for this. The user or the scribe may use permitAND and permitOR to further customize this filtering, even dynamically if they wish to.

Verbosity is used to select keys from the log item's payload. Each LogItem instance describes what keys should be retained for each Verbosity level. Use the payloadObject utility for extracting the keys that should be written.

Scribes provide a finalizer IO action (scribeFinalizer) that is meant to synchronously flush any remaining writes and clean up any resources acquired when the scribe was created. Internally, katip keeps a buffer for each scribe's writes. When closeScribe or closeScribes is called, that buffer stops accepting new log messages and after the last item in its buffer is sent to liPush, calls the finalizer. Thus, when the finalizer returns, katip can assume that all resources are cleaned up and all log messages are durably written.

While katip internally buffers messages per ScribeSettings, it sends them one at a time to the scribe. Depending on the scribe itself, it may make sense for that scribe to keep its own internal buffer to batch-send logs if writing items one at a time is not efficient. The scribe implementer must be sure that on finalization, all writes are committed synchronously.

Signature of a function passed to Scribe constructor and mkScribe* functions that decides which messages to be logged. Typically filters based on Severity, but can be combined with other, custom logic with permitAND and permitOR

permitAND :: PermitFunc -> PermitFunc -> PermitFunc

AND together 2 permit functions

permitOR :: PermitFunc -> PermitFunc -> PermitFunc

OR together 2 permit functions

data Scribe

Constructors

Scribe

Fields

liPush :: forall a. LogItem a => Item a -> IO ()

How do we write an item to the scribe's output?

scribeFinalizer :: IO ()

Provide a blocking finalizer to call when your scribe is removed. All pending writes should be flushed synchronously. If this is not relevant to your scribe, return () is fine.

scribePermitItem :: PermitFunc

Provide a filtering function to allow the item to be logged, or not. It can check Severity or some string in item's body. The initial value of this is usually created from permitItem. Scribes and users can customize this by ANDing or ORing onto the default with permitAND or permitOR

whenM :: Monad m => m Bool -> m () -> m ()

data ScribeHandle

Constructors

ScribeHandle

Fields

shScribe :: Scribe

shChan :: TBQueue WorkerMessage

data WorkerMessage where

Constructors

NewItem :: LogItem a => Item a -> WorkerMessage

PoisonPill :: WorkerMessage

permitItem :: Monad m => Severity -> Item a -> m Bool

Should this item be logged given the user's maximum severity? Most new scribes will use this as a base for their PermitFunc

data LogEnv

Constructors

LogEnv

Fields

_logEnvHost :: HostName

_logEnvPid :: ProcessID

_logEnvApp :: Namespace

Name of application. This will typically never change. This field gets prepended to the namespace of your individual log messages. For example, if your app is MyApp and you write a log using "logItem" and the namespace WebServer, the final namespace will be MyApp.WebServer

_logEnvEnv :: Environment

_logEnvTimer :: IO UTCTime

Action to fetch the timestamp. You can use something like AutoUpdate for high volume logs but note that this may cause some output forms to display logs out of order. Alternatively, you could just use getCurrentTime.

_logEnvScribes :: Map Text ScribeHandle

logEnvTimer :: Lens' LogEnv (IO UTCTime)

logEnvScribes :: Lens' LogEnv (Map Text ScribeHandle)

logEnvPid :: Lens' LogEnv ProcessID

logEnvHost :: Lens' LogEnv HostName

logEnvEnv :: Lens' LogEnv Environment

logEnvApp :: Lens' LogEnv Namespace

initLogEnv

Arguments

:: Namespace
A base namespace for this application

-> Environment
Current run environment (e.g. prod vs. devel)

-> IO LogEnv

Create a reasonable default InitLogEnv. Uses an AutoUpdate which updates the timer every 1ms. If you need even more timestamp precision at the cost of performance, consider setting _logEnvTimer with getCurrentTime.

registerScribe

Arguments

:: Text
Name the scribe

-> Scribe

-> ScribeSettings

-> LogEnv

-> IO LogEnv

Add a scribe to the list. All future log calls will go to this scribe in addition to the others. Writes will be buffered per the ScribeSettings to prevent slow scribes from slowing down logging. Writes will be dropped if the buffer fills.

spawnScribeWorker :: Scribe -> TBQueue WorkerMessage -> IO (Async ())

data ScribeSettings

Constructors

ScribeSettings

Fields

_scribeBufferSize :: Int

scribeBufferSize :: Lens' ScribeSettings Int

defaultScribeSettings :: ScribeSettings

Reasonable defaults for a scribe. Buffer size of 4096.

unregisterScribe

Arguments

:: Text
Name of the scribe

-> LogEnv

-> LogEnv

Remove a scribe from the environment. This does not finalize the scribe. This mainly only makes sense to use with something like MonadReader's local function to temporarily disavow a single logger for a block of code.

clearScribes :: LogEnv -> LogEnv

Unregister all scribes. Note that this is not for closing or finalizing scribes, use closeScribes for that. This mainly only makes sense to use with something like MonadReader's local function to temporarily disavow any loggers for a block of code.

closeScribe

Arguments

:: Text
Name of the scribe

-> LogEnv

-> IO LogEnv

Finalize a scribe. The scribe is removed from the environment, its finalizer is called so that it can never be written to again and all pending writes are flushed. Note that this will throw any exceptions yoru finalizer will throw, and that LogEnv is immutable, so it will not be removed in that case.

closeScribes :: LogEnv -> IO LogEnv

Call this at the end of your program. This is a blocking call that stop writing to a scribe's queue, waits for the queue to empty, finalizes each scribe in the log environment and then removes it. Finalizers are all run even if one of them throws, but the exception will be re-thrown at the end.

class MonadIO m => Katip m where

Monads where katip logging actions can be performed. Katip is the most basic logging monad. You will typically use this directly if you either don't want to use namespaces/contexts heavily or if you want to pass in specific contexts and/or namespaces at each log site.

For something more powerful, look at the docs for KatipContext, which keeps a namespace and merged context. You can write simple functions that add additional namespacing and merges additional context on the fly.

localLogEnv was added to allow for lexically-scoped modifications of the log env that are reverted when the supplied monad completes. katipNoLogging, for example, uses this to temporarily pause log outputs.

Methods

getLogEnv :: m LogEnv

localLogEnv :: (LogEnv -> LogEnv) -> m a -> m a

newtype KatipT m a

A concrete monad you can use to run logging actions. Use this if you prefer an explicit monad transformer stack and adding layers as opposed to implementing Katip for your monad.

Constructors

KatipT

Fields

unKatipT :: ReaderT LogEnv m a

runKatipT :: LogEnv -> KatipT m a -> m a

Execute KatipT on a log env.

katipNoLogging :: Katip m => m a -> m a

Disable all scribes for the given monadic action, then restore them afterwards. Works in any Katip monad.

logItem :: (Applicative m, LogItem a, Katip m) => a -> Namespace -> Maybe Loc -> Severity -> LogStr -> m ()

Log with everything, including a source code location. This is very low level and you typically can use logT in its place.

logKatipItem :: (Applicative m, LogItem a, Katip m) => Item a -> m ()

Log already constructed Item. This is the lowest level function that other log* functions use. It can be useful when implementing centralised logging services.

tryWriteTBQueue

Arguments

:: TBQueue a

-> a

-> STM Bool
Did we write?

logF

Arguments

:: (Applicative m, LogItem a, Katip m)

=> a
Contextual payload for the log

-> Namespace
Specific namespace of the message.

-> Severity
Severity of the message

-> LogStr
The log message

-> m ()

Log with full context, but without any code location.

logException

Arguments

:: (Katip m, LogItem a, MonadCatch m, Applicative m)

=> a
Log context

-> Namespace
Namespace

-> Severity
Severity

-> m b
Main action being run

-> m b

Perform an action while logging any exceptions that may occur. Inspired by onException.

>>> > logException () mempty ErrorS (error "foo")

logMsg :: (Applicative m, Katip m) => Namespace -> Severity -> LogStr -> m ()

Log a message without any payload/context or code location.

liftLoc :: Loc -> Q Exp

Lift a location into an Exp.

getLoc :: HasCallStack => Maybe Loc

For use when you want to include location in your logs. This will fill the 'Maybe Loc' gap in logF of this module, and relies on implicit callstacks when available (GHC > 7.8).

getLocTH :: ExpQ

logT :: ExpQ

Loc-tagged logging when using template-haskell.

$(logT) obj mempty InfoS "Hello world"

logLoc :: (Applicative m, LogItem a, Katip m, HasCallStack) => a -> Namespace -> Severity -> LogStr -> m ()

Loc-tagged logging using Stack when available.

This function does not require template-haskell as it automatically uses implicit-callstacks when the code is compiled using GHC > 7.8. Using an older version of the compiler will result in the emission of a log line without any location information, so be aware of it. Users using GHC <= 7.8 may want to use the template-haskell function logT for maximum compatibility.

logLoc obj mempty InfoS "Hello world"

locationToString :: Loc -> String